backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
PyJWT>=2.8.0
django-cors-headers>=4.3.0
python-decouple>=3.8
pytest-django>=4.5.0
pytest>=7.4.0
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0

# PostgreSQL support
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# Redis support
django-redis>=5.4.0
redis>=5.0.0

# Production server
gunicorn>=21.2.0
